Finished swiftUI integration
