+++ PennMobile/About/Privacy/PrivacyTableViewController.swift
@@ -17,8 +17,6 @@ class PrivacyViewController: GenericTableViewController, ShowsAlert, IndicatorEn
     
     let displayedPrefs = PrivacyOption.visibleOptions
     
-    var cancellable: Any? = nil
-    
     override func viewDidLoad() {
         super.viewDidLoad()
         
@@ -35,26 +33,6 @@ class PrivacyViewController: GenericTableViewController, ShowsAlert, IndicatorEn
     override func viewWillAppear(_ animated: Bool) {
         super.viewDidAppear(animated)
         tableView.reloadData()
-        
-        if #available(iOS 13, *) {
-            let delegate = PrivacyPermissionDelegate()
-            let vc = UIHostingController(rootView: PrivacyPermissionView(delegate: delegate))
-            present(vc, animated: true)
-
-            self.cancellable = delegate.objectDidChange.sink { (delegate) in
-                if let decision = delegate.userDecision {
-                    switch decision {
-                    case .affirmative: print("CONSENT GIVEN")
-                    case .negative: print("NO CONSENT GIVEN")
-                    case .moreInfo: print("INFO REQUESTED")
-                    case .close: print("CLOSE VIEW")
-                    }
-                    //vc.dismiss(animated: true, completion: nil)
-                }
-            }
-        } else {
-            // Fallback on earlier versions
-        }
     }
 }
